FEM: task panels, move them into the view provider modules
