+++ src/Mod/Fem/PyGui/_ViewProviderFemMeshGroup.py
@@ -29,7 +29,11 @@ __url__ = "http://www.freecadweb.org"
 
 import FreeCAD
 import FreeCADGui
-from pivy import coin
+
+
+# for the panel
+from PySide import QtCore
+from PySide import QtGui
 
 
 class _ViewProviderFemMeshGroup:
@@ -41,6 +45,7 @@ class _ViewProviderFemMeshGroup:
         return ":/icons/fem-femmesh-from-shape.svg"
 
     def attach(self, vobj):
+        from pivy import coin
         self.ViewObject = vobj
         self.Object = vobj.Object
         self.standard = coin.SoGroup()
@@ -64,8 +69,7 @@ class _ViewProviderFemMeshGroup:
             if o.isDerivedFrom("Fem::FemMeshObject"):
                 o.ViewObject.hide()
         # show task panel
-        import PyGui._TaskPanelFemMeshGroup
-        taskd = PyGui._TaskPanelFemMeshGroup._TaskPanelFemMeshGroup(self.Object)
+        taskd = _TaskPanelFemMeshGroup(self.Object)
         taskd.obj = vobj.Object
         FreeCADGui.Control.showDialog(taskd)
         return True
@@ -87,3 +91,195 @@ class _ViewProviderFemMeshGroup:
 
     def __setstate__(self, state):
         return None
+
+
+class _TaskPanelFemMeshGroup:
+    '''The TaskPanel for editing References property of FemMeshGroup objects'''
+
+    def __init__(self, obj):
+        FreeCADGui.Selection.clearSelection()
+        self.sel_server = None
+        self.obj = obj
+        self.selection_mode_solid = False
+        self.selection_mode_std_print_message = "Select Faces, Edges and Vertices by single click on them to add them to the list."
+        self.selection_mode_solid_print_message = "Select Solids by single click on a Face or Edge which belongs to the Solid, to add the Solid to the list."
+        self.obj_notvisible = []
+
+        self.form = FreeCADGui.PySideUic.loadUi(FreeCAD.getHomePath() + "Mod/Fem/PyGui/TaskPanelFemMeshGroup.ui")
+        QtCore.QObject.connect(self.form.rb_name, QtCore.SIGNAL("toggled(bool)"), self.choose_exportidentifier_name)
+        QtCore.QObject.connect(self.form.rb_label, QtCore.SIGNAL("toggled(bool)"), self.choose_exportidentifier_label)
+        QtCore.QObject.connect(self.form.rb_standard, QtCore.SIGNAL("toggled(bool)"), self.choose_selection_mode_standard)
+        QtCore.QObject.connect(self.form.rb_solid, QtCore.SIGNAL("toggled(bool)"), self.choose_selection_mode_solid)
+        QtCore.QObject.connect(self.form.pushButton_Reference, QtCore.SIGNAL("clicked()"), self.add_references)
+        self.form.list_References.itemSelectionChanged.connect(self.select_clicked_reference_shape)
+        self.form.list_References.setContextMenuPolicy(QtCore.Qt.CustomContextMenu)
+        self.form.list_References.connect(self.form.list_References, QtCore.SIGNAL("customContextMenuRequested(QPoint)"), self.references_list_right_clicked)
+
+        self.get_meshgroup_props()
+        self.update()
+
+    def accept(self):
+        self.setback_listobj_visibility()
+        self.set_meshgroup_props()
+        if self.sel_server:
+            FreeCADGui.Selection.removeObserver(self.sel_server)
+        FreeCADGui.ActiveDocument.resetEdit()
+        FreeCAD.ActiveDocument.recompute()
+        return True
+
+    def reject(self):
+        self.setback_listobj_visibility()
+        if self.sel_server:
+            FreeCADGui.Selection.removeObserver(self.sel_server)
+        FreeCADGui.ActiveDocument.resetEdit()
+        return True
+
+    def get_meshgroup_props(self):
+        self.use_label = self.obj.UseLabel
+        self.references = []
+        if self.obj.References:
+            self.tuplereferences = self.obj.References
+            self.get_references()
+
+    def set_meshgroup_props(self):
+        self.obj.References = self.references
+        self.obj.UseLabel = self.use_label
+
+    def update(self):
+        'fills the widgets'
+        self.form.rb_name.setChecked(not self.use_label)
+        self.form.rb_label.setChecked(self.use_label)
+        self.rebuild_list_References()
+
+    def choose_exportidentifier_name(self, state):
+        self.use_label = not state
+
+    def choose_exportidentifier_label(self, state):
+        self.use_label = state
+
+    def choose_selection_mode_standard(self, state):
+        self.selection_mode_solid = not state
+        if self.sel_server and not self.selection_mode_solid:
+            print(self.selection_mode_std_print_message)
+
+    def choose_selection_mode_solid(self, state):
+        self.selection_mode_solid = state
+        if self.sel_server and self.selection_mode_solid:
+            print(self.selection_mode_solid_print_message)
+
+    def get_references(self):
+        for ref in self.tuplereferences:
+            for elem in ref[1]:
+                self.references.append((ref[0], elem))
+
+    def references_list_right_clicked(self, QPos):
+        self.form.contextMenu = QtGui.QMenu()
+        menu_item = self.form.contextMenu.addAction("Remove Reference")
+        if not self.references:
+            menu_item.setDisabled(True)
+        self.form.connect(menu_item, QtCore.SIGNAL("triggered()"), self.remove_reference)
+        parentPosition = self.form.list_References.mapToGlobal(QtCore.QPoint(0, 0))
+        self.form.contextMenu.move(parentPosition + QPos)
+        self.form.contextMenu.show()
+
+    def remove_reference(self):
+        if not self.references:
+            return
+        currentItemName = str(self.form.list_References.currentItem().text())
+        for ref in self.references:
+            refname_to_compare_listentry = ref[0].Name + ':' + ref[1]
+            if refname_to_compare_listentry == currentItemName:
+                self.references.remove(ref)
+        self.rebuild_list_References()
+
+    def add_references(self):
+        '''Called if Button add_reference is triggered'''
+        # in constraints EditTaskPanel the selection is active as soon as the taskpanel is open
+        # here the addReference button EditTaskPanel has to be triggered to start selection mode
+        self.setback_listobj_visibility()
+        FreeCADGui.Selection.clearSelection()
+        # start SelectionObserver and parse the function to add the References to the widget
+        if self.selection_mode_solid:  # print message on button click
+            print_message = self.selection_mode_solid_print_message
+        else:
+            print_message = self.selection_mode_std_print_message
+        if not self.sel_server:
+            # if we do not check, we would start a new SelectionObserver on every click on addReference button
+            # but close only one SelectionObserver on leaving the task panel
+            from . import FemSelectionObserver
+            self.sel_server = FemSelectionObserver.FemSelectionObserver(self.selectionParser, print_message)
+
+    def selectionParser(self, selection):
+        print('selection: ', selection[0].Shape.ShapeType, '  ', selection[0].Name, '  ', selection[1])
+        if hasattr(selection[0], "Shape") and selection[1]:
+            elt = selection[0].Shape.getElement(selection[1])
+            if self.selection_mode_solid:
+                # in solid selection mode use edges and faces for selection of a solid
+                solid_to_add = None
+                if elt.ShapeType == 'Edge':
+                    found_edge = False
+                    for i, s in enumerate(selection[0].Shape.Solids):
+                        for e in s.Edges:
+                            if elt.isSame(e):
+                                if not found_edge:
+                                    solid_to_add = str(i + 1)
+                                else:
+                                    FreeCAD.Console.PrintMessage('Edge belongs to more than one solid\n')
+                                    solid_to_add = None
+                                found_edge = True
+                elif elt.ShapeType == 'Face':
+                    found_face = False
+                    for i, s in enumerate(selection[0].Shape.Solids):
+                        for e in s.Faces:
+                            if elt.isSame(e):
+                                if not found_face:
+                                    solid_to_add = str(i + 1)
+                                else:
+                                    FreeCAD.Console.PrintMessage('Face belongs to more than one solid\n')
+                                    solid_to_add = None
+                                found_edge = True
+                if solid_to_add:
+                    selection = (selection[0], 'Solid' + solid_to_add)
+                    print('selection element changed to Solid: ', selection[0].Shape.ShapeType, '  ', selection[0].Name, '  ', selection[1])
+                else:
+                    return
+            if selection not in self.references:
+                self.references.append(selection)
+                self.rebuild_list_References()
+            else:
+                FreeCAD.Console.PrintMessage(selection[0].Name + ' --> ' + selection[1] + ' is in reference list already!\n')
+
+    def rebuild_list_References(self):
+        self.form.list_References.clear()
+        items = []
+        for ref in self.references:
+            item_name = ref[0].Name + ':' + ref[1]
+            items.append(item_name)
+        for listItemName in sorted(items):
+            self.form.list_References.addItem(listItemName)
+
+    def select_clicked_reference_shape(self):
+        self.setback_listobj_visibility()
+        if self.sel_server:
+            FreeCADGui.Selection.removeObserver(self.sel_server)
+            self.sel_server = None
+        if not self.sel_server:
+            if not self.references:
+                return
+            currentItemName = str(self.form.list_References.currentItem().text())
+            for ref in self.references:
+                refname_to_compare_listentry = ref[0].Name + ':' + ref[1]
+                if refname_to_compare_listentry == currentItemName:
+                    # print( 'found: shape: ' + ref[0].Name + ' element: ' + ref[1])
+                    if not ref[0].ViewObject.Visibility:
+                        self.obj_notvisible.append(ref[0])
+                        ref[0].ViewObject.Visibility = True
+                    FreeCADGui.Selection.clearSelection()
+                    FreeCADGui.Selection.addSelection(ref[0], ref[1])
+
+    def setback_listobj_visibility(self):
+        '''set back Visibility of the list objects
+        '''
+        for obj in self.obj_notvisible:
+            obj.ViewObject.Visibility = False
+        self.obj_notvisible = []
